refactor: move logic to external files
