+++ src/components/nSelect/logic/index.js
@@ -1,3 +1,5 @@
+import nuiHandler from './nuiHandler'
+import styleHandler from './styleHandler'
 import indexHandler from './indexHandler'
 import eventsHandler from './eventsHandler'
 import keyboardHandler from './keyboardHandler'
@@ -5,5 +7,19 @@ import validationHandler from './validationHandler'
 import multiSelectHandler from './multiSelectHandler'
 
 export default {
-    mixins: [eventsHandler, keyboardHandler, indexHandler, validationHandler, multiSelectHandler]
+  props: {
+    optionHoverBackgroundColor: {
+      type: String,
+      default: ''
+    }
+  },
+  mixins: [
+    styleHandler,
+    eventsHandler,
+    keyboardHandler,
+    indexHandler,
+    validationHandler,
+    multiSelectHandler,
+    nuiHandler
+  ]
 }
